角色编辑，和异常处理

+++ src/main/java/cn/merryyou/logback/domain/SysMenu.java
@@ -80,7 +80,7 @@ public class  SysMenu implements Serializable {
      */
     private String permission;
 
-    private Byte menuType;
+    private int menuType;
 
 
     @Override
